Tag tweets with hashtags from within the message
Via the twitter API (Zip not supported yet) you can now have
the messages tagged with the hashtag from the tweet. Optionally
you can also have them tagged with "twitter" (both must be on
for this to work).

Updated the HTML to show what attributes to set to true.

This change allows you to search more easily for tweets via
the seach box with tag:<tag> or "tag:twitter" for all tweets.

Change-Id: I71aeea5e3340774d7341b297e0815974e66c309f